Rename all components to match NextJS naming standards
